requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
python-jose[cryptography]
passlib[bcrypt]
apscheduler
minio
pillow